chat/requirements.txt
Flask-SQLAlchemy>=3.1
pyjwt[crypto]
bcrypt>=3.2.0
mysql-connector-python>=8.0
requests
